requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
pytest==8.4.2
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2